+++ index.js
@@ -1,11 +1,13 @@
 const httpClient = require('./src/httpClient');
 const AccessRecordsApi = require('./src/accessRecordsApi');
 const UserPermissionsApi = require('./src/userPermissionsApi');
+const UsersApi = require('./src/usersApi');
 const ServiceClientsApi = require('./src/serviceClientsApi');
 const ResourcesApi = require('./src/resourcesApi');
 const AccountsApi = require('./src/accountsApi');
 const RolesApi = require('./src/rolesApi');
 const ConnectionsApi = require('./src/connectionsApi');
+const TenantsApi = require('./src/tenantsApi');
 
 class AuthressClient {
   constructor(settings, tokenProvider) {
@@ -16,10 +18,12 @@ class AuthressClient {
     this.accessRecords = new AccessRecordsApi(this.httpClient);
     this.serviceClients = new ServiceClientsApi(this.httpClient);
     this.userPermissions = new UserPermissionsApi(this.httpClient);
+    this.users = new UsersApi(this.httpClient);
     this.resources = new ResourcesApi(this.httpClient);
     this.accounts = new AccountsApi(this.httpClient);
     this.roles = new RolesApi(this.httpClient);
     this.connections = new ConnectionsApi(this.httpClient);
+    this.tenants = new TenantsApi(this.httpClient);
   }
 
   setToken(token) {