+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This is the changelog for [Authress SDK](readme.md).
 * Add EdDSA support for `tokenVerifier()` class
 * Set the service client authorization request type to be `oauth-authz-req+jwt`
 * Handle malformed baseUrls in `httpClient`.
+* Allow specifying the authress custom domain for service client machine to machine authentication.
+* Add `users.getUser(userId)` api method.
+* Add `connections` API to the SDK.
 
 ## 1.1 ##
 * Migrated to Github Actions